examples: drop the run-example.ts indirection, call runScript directly
Two layers of trivial wrapping (apps/os's shared runExample/
runScriptEnvelope, re-wrapped by apps/mobile's runMobileExample) around
one line of capabilityHost.runScript(...). Reverted apps/os/src/itx/
run-example.ts back to its original apps/os/e2e/test-support/ home
(unchanged from before this task); apps/mobile's Examples screen and
its e2e now just look the example up and call runScript directly, with
the envelope inlined — a little repetition beats an abstraction for
one call site's worth of logic.
